office_emp_proj_with_auth/accounts/views.py
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response
from rest_framework_simplejwt.tokens import RefreshToken

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def social_login_token(request):
    """
    Call this after logging in via django-allauth (session cookie present).
    Returns JWT refresh & access tokens for the current user.
    """
    user = request.user
    refresh = RefreshToken.for_user(user)
    return Response({'refresh': str(refresh), 'access': str(refresh.access_token)})
