office_emp_proj_with_auth/requirements.txt
Django>=5.1
djangorestframework
djangorestframework-simplejwt
django-allauth
